requirements.txt
Bootstrap-Flask==2.0.2
Flask==2.1.2
Flask-Migrate==3.1.0
flask-msearch==0.2.9.2
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
SQLAlchemy==1.4.39
